LPS-65385 Move DestinationNames to com.liferay.portal.configuration.cluster.constants
