+++ modules/apps/foundation/portal-configuration/portal-configuration-cluster/src/main/java/com/liferay/portal/configuration/cluster/internal/messaging/ConfigurationMessageListener.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.configuration.cluster.internal.messaging;
 
+import com.liferay.portal.configuration.cluster.constants.ConfigurationClusterDestinationNames;
 import com.liferay.portal.configuration.cluster.internal.ConfigurationThreadLocal;
-import com.liferay.portal.configuration.cluster.messaging.DestinationNames;
 import com.liferay.portal.configuration.persistence.ReloadablePersistenceManager;
 import com.liferay.portal.kernel.messaging.BaseMessageListener;
 import com.liferay.portal.kernel.messaging.Destination;
@@ -37,7 +37,7 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(
 	immediate = true,
-	property = {"destination.name=" + DestinationNames.CONFIGURATION},
+	property = {"destination.name=" + ConfigurationClusterDestinationNames.CONFIGURATION},
 	service = MessageListener.class
 )
 public class ConfigurationMessageListener extends BaseMessageListener {
@@ -113,7 +113,7 @@ public class ConfigurationMessageListener extends BaseMessageListener {
 	}
 
 	@Reference(
-		target = "(destination.name=" + DestinationNames.CONFIGURATION + ")",
+		target = "(destination.name=" + ConfigurationClusterDestinationNames.CONFIGURATION + ")",
 		unbind = "-"
 	)
 	protected void setDestination(Destination destination) {
